refactor into regex package
